Unite `client` and `server` types
